GameSearch feature works in a simulated database with "genre" filter
When entering to <http://localhost:8080/gameSearch> with query parameter "name" a search is made in a postgres database.
A table in schema power_up named game, with columns name, summary and avg_score in postgres with user=postgres and password=paw is needed.
A non-required parameter "genre" can be added to filter the search.
The Publisher or Developer filters don't work