Game genre, publisher, developer and platform in the game description make a search with respective filter when clicked
Replace in game.jsp the values of the genre,publisher, developer and platform with a button having the same value. Button looks like a hyperlink.
When clicked a js function is call with the name of the filter and the value and a new search in /search page is made with the new parameters.