Release Date now shows real release date in the game details page.
Platform are now stored in a map with <String, DateTime>, with String being the platform name and DateTime the Release Date for that platform.
Date is showed in its full length in the gameDetails page. It should be adapted for a more user friendly number.
Test for the releaseDate of Platform was added.